+++ vidgear/tests/network_tests/asyncio_tests/test_netgear_async.py
@@ -27,9 +27,11 @@ import platform
 import numpy as np
 import pytest
 import asyncio
+import functools
 import logging as log
 import tempfile
 
+from async_timeout import timeout
 from vidgear.gears.asyncio import NetGear_Async
 from vidgear.gears.asyncio.helper import logger_handler
 
@@ -48,6 +50,19 @@ def return_testvideo_path():
         tempfile.gettempdir()
     )
     return os.path.abspath(path)
+
+
+def with_timeout(t):
+    def wrapper(corofunc):
+        @functools.wraps(corofunc)
+        async def run(*args, **kwargs):
+            with timeout(t):
+                print(kwargs)
+                return await corofunc(*args, **kwargs)
+
+        return run
+
+    return wrapper
 
 
 # Create a async frame generator as custom source
@@ -88,6 +103,8 @@ def event_loop():
 
 
 @pytest.mark.asyncio
+@pytest.mark.xfail(raises=asyncio.TimeoutError)
+@with_timeout(5)
 @pytest.mark.parametrize(
     "pattern",
     [0, 2, 3, 4],
@@ -124,6 +141,8 @@ test_data_class = [
 
 
 @pytest.mark.asyncio
+@pytest.mark.xfail(raises=asyncio.TimeoutError)
+@with_timeout(5)
 @pytest.mark.parametrize("generator, result", test_data_class)
 async def test_netgear_async_custom_server_generator(generator, result):
     try:
@@ -145,6 +164,8 @@ async def test_netgear_async_custom_server_generator(generator, result):
 
 
 @pytest.mark.asyncio
+@pytest.mark.xfail(raises=asyncio.TimeoutError)
+@with_timeout(5)
 @pytest.mark.parametrize("address, port", [("172.31.11.15.77", "5555"), (None, "5555")])
 async def test_netgear_async_addresses(address, port):
     try:
@@ -178,7 +199,8 @@ async def test_netgear_async_addresses(address, port):
 
 
 @pytest.mark.asyncio
-@pytest.mark.xfail(raises=ValueError)
+@pytest.mark.xfail(raises=(ValueError, asyncio.TimeoutError))
+@with_timeout(5)
 async def test_netgear_async_recv_generator():
     # define and launch server
     server = NetGear_Async(source=return_testvideo_path(), logging=True)