+++ vidgear/tests/writer_tests/test_compression_mode.py
@@ -28,6 +28,7 @@ import logging as log
 import platform
 import tempfile
 import subprocess
+import timeout_decorator
 from six import string_types
 
 from vidgear.gears import CamGear, WriteGear
@@ -70,6 +71,17 @@ def return_testvideo_path():
     return os.path.abspath(path)
 
 
+def remove_file_safe(path):
+    """
+    Remove file safely
+    """
+    try:
+        if path and os.path.isfile(os.path.abspath(path)):
+            os.remove(path)
+    except Exception as e:
+        logger.exception(e)
+
+
 def getFrameRate(path):
     """
     Returns framerate of video(at path provided) using FFmpeg
@@ -85,7 +97,8 @@ def getFrameRate(path):
     return float(match_dict["fps"])
 
 
-@pytest.mark.xfail(raises=AssertionError)
+@pytest.mark.xfail(raises=(AssertionError, StopIteration))
+@timeout_decorator.timeout(5, timeout_exception=StopIteration)
 @pytest.mark.parametrize("c_ffmpeg", [return_static_ffmpeg(), "wrong_path"])
 def test_input_framerate(c_ffmpeg):
     """
@@ -113,9 +126,11 @@ def test_input_framerate(c_ffmpeg):
     writer.close()
     output_video_framerate = getFrameRate(os.path.abspath("Output_tif.mp4"))
     assert test_video_framerate == output_video_framerate
-    os.remove(os.path.abspath("Output_tif.mp4"))
+    remove_file_safe("Output_tif.mp4")
 
 
+@pytest.mark.xfail(raises=StopIteration)
+@timeout_decorator.timeout(5, timeout_exception=StopIteration)
 @pytest.mark.parametrize(
     "conversion", ["COLOR_BGR2GRAY", "COLOR_BGR2INVALID", "COLOR_BGR2BGRA"]
 )
@@ -184,11 +199,11 @@ def test_write(conversion):
         else:
             logger.exception(str(e))
     finally:
-        if os.path.isfile(os.path.abspath("Output_tw.mp4")):
-            os.remove(os.path.abspath("Output_tw.mp4"))
+        remove_file_safe("Output_tw.mp4")
 
 
-@pytest.mark.xfail(raises=AssertionError)
+@pytest.mark.xfail(raises=(AssertionError, StopIteration))
+@timeout_decorator.timeout(5, timeout_exception=StopIteration)
 def test_output_dimensions():
     """
     Testing "-output_dimensions" special parameter provided by WriteGear(in Compression Mode)
@@ -225,7 +240,7 @@ def test_output_dimensions():
     assert output_dim[0] == 640 and output_dim[1] == 480
     output.release()
 
-    os.remove(os.path.abspath("Output_tod.mp4"))
+    remove_file_safe("Output_tod.mp4")
 
 
 test_data_class = [
@@ -242,6 +257,8 @@ test_data_class = [
 ]
 
 
+@pytest.mark.xfail(raises=StopIteration)
+@timeout_decorator.timeout(5, timeout_exception=StopIteration)
 @pytest.mark.parametrize("f_name, c_ffmpeg, output_params, result", test_data_class)
 def test_WriteGear_compression(f_name, c_ffmpeg, output_params, result):
     """
@@ -259,8 +276,7 @@ def test_WriteGear_compression(f_name, c_ffmpeg, output_params, result):
             writer.write(frame)
         stream.release()
         writer.close()
-        if f_name and os.path.isfile(os.path.abspath(f_name)):
-            os.remove(os.path.abspath(f_name))
+        remove_file_safe(f_name)
     except Exception as e:
         if result:
             pytest.fail(str(e))
@@ -303,6 +319,8 @@ def test_WriteGear_compression(f_name, c_ffmpeg, output_params, result):
         ),
     ],
 )
+@pytest.mark.xfail(raises=StopIteration)
+@timeout_decorator.timeout(5, timeout_exception=StopIteration)
 def test_WriteGear_customFFmpeg(ffmpeg_command_to_save_audio, logging, output_params):
     """
     Testing WriteGear Compression-Mode(FFmpeg) custom FFmpeg Pipeline by seperating audio from video