+++ vidgear/tests/writer_tests/test_non_compression_mode.py
@@ -25,6 +25,7 @@ import pytest
 import logging as log
 import platform
 import tempfile
+import timeout_decorator
 
 from six import string_types
 from vidgear.gears import WriteGear
@@ -67,7 +68,19 @@ def return_testvideo_path():
     return os.path.abspath(path)
 
 
-@pytest.mark.xfail(raises=AssertionError)
+def remove_file_safe(path):
+    """
+    Remove file safely
+    """
+    try:
+        if path and os.path.isfile(os.path.abspath(path)):
+            os.remove(path)
+    except Exception as e:
+        logger.exception(e)
+
+
+@pytest.mark.xfail(raises=(AssertionError, StopIteration))
+@timeout_decorator.timeout(5, timeout_exception=StopIteration)
 @pytest.mark.parametrize("conversion", ["COLOR_BGR2GRAY", "COLOR_BGR2YUV"])
 def test_write(conversion):
     """
@@ -109,7 +122,7 @@ def test_write(conversion):
         logger.debug("Result: {}".format(result))
         for i in ["Error", "Invalid", "error", "invalid"]:
             assert not (i in result)
-    os.remove(os.path.abspath("Output_twc.avi"))
+    remove_file_safe("Output_twc.avi")
 
 
 test_data_class = [
@@ -127,7 +140,8 @@ test_data_class = [
     ),
 ]
 
-
+@pytest.mark.xfail(raises=StopIteration)
+@timeout_decorator.timeout(5, timeout_exception=StopIteration)
 @pytest.mark.parametrize("f_name, output_params, result", test_data_class)
 def test_WriteGear_compression(f_name, output_params, result):
     """
@@ -148,8 +162,7 @@ def test_WriteGear_compression(f_name, output_params, result):
             writer.write(frame)
         stream.release()
         writer.close()
-        if f_name and os.path.isfile(os.path.abspath(f_name)):
-            os.remove(os.path.abspath(f_name))
+        remove_file_safe(f_name)
     except Exception as e:
         if result:
             pytest.fail(str(e))